requirements.txt
robotframework==6.0.2
influxdb==5.3.1
requests==2.28.1
pytest==7.2.0
mypy==1.0.0